WIP:bumping dependencies up to more recent version of ODL, compiles but doesn't start up

+++ loopremover/implementation/src/main/java/org/opendaylight/l2switch/loopremover/LoopRemoverProvider.java
@@ -7,7 +7,7 @@
  */
 package org.opendaylight.l2switch.loopremover;
 
-import org.opendaylight.controller.md.sal.binding.api.DataBroker;
+import org.opendaylight.mdsal.binding.api.DataBroker;
 import org.opendaylight.l2switch.loopremover.flow.InitialFlowWriter;
 import org.opendaylight.l2switch.loopremover.topology.NetworkGraphImpl;
 import org.opendaylight.l2switch.loopremover.topology.NetworkGraphService;
@@ -17,6 +17,9 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.packet.loop.remover.config.
 import org.opendaylight.yangtools.concepts.Registration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.opendaylight.yangtools.yang.common.Uint8;
+import org.opendaylight.yangtools.yang.common.Uint16;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 public class LoopRemoverProvider {
 
@@ -43,17 +46,17 @@ public class LoopRemoverProvider {
         if (loopRemoverConfig.isIsInstallLldpFlow()) {
             LOG.info("LoopRemover will install an lldp flow");
             InitialFlowWriter initialFlowWriter = new InitialFlowWriter(salFlowService);
-            initialFlowWriter.setFlowTableId(loopRemoverConfig.getLldpFlowTableId());
-            initialFlowWriter.setFlowPriority(loopRemoverConfig.getLldpFlowPriority());
-            initialFlowWriter.setFlowIdleTimeout(loopRemoverConfig.getLldpFlowIdleTimeout());
-            initialFlowWriter.setFlowHardTimeout(loopRemoverConfig.getLldpFlowHardTimeout());
+            initialFlowWriter.setFlowTableId(Uint16.valueOf(loopRemoverConfig.getLldpFlowTableId()).shortValue());
+            initialFlowWriter.setFlowPriority(loopRemoverConfig.getLldpFlowPriority().intValue());
+            initialFlowWriter.setFlowIdleTimeout(loopRemoverConfig.getLldpFlowIdleTimeout().intValue());
+            initialFlowWriter.setFlowHardTimeout(loopRemoverConfig.getLldpFlowHardTimeout().intValue());
             topoNodeListnerReg = initialFlowWriter.registerAsDataChangeListener(dataService);
         }
 
         // Register Topology DataChangeListener
         NetworkGraphService networkGraphService = new NetworkGraphImpl();
         this.topologyLinkDataChangeHandler = new TopologyLinkDataChangeHandler(dataService, networkGraphService);
-        topologyLinkDataChangeHandler.setGraphRefreshDelay(loopRemoverConfig.getGraphRefreshDelay());
+        topologyLinkDataChangeHandler.setGraphRefreshDelay(loopRemoverConfig.getGraphRefreshDelay().longValue());
         topologyLinkDataChangeHandler.setTopologyId(loopRemoverConfig.getTopologyId());
         listenerRegistration = topologyLinkDataChangeHandler.registerAsDataChangeListener();
